# SIMD: Extend the enabled targets for Google Highway quicksort