@seiko2plus 2024-01-10
The current version of Google Highway seems to break the build on both PowerPC and IBMZ.