@Mousius 2024-01-10
> The current version of Google Highway seems to break the build on both PowerPC and IBMZ.

Did you try the latest? I know there's been a lot of changes to 16-bit and @jan-wassenberg was planning another release 😸