@Mousius 2024-01-11
> The [s390 issue](https://github.com/numpy/numpy/actions/runs/7477479830/job/20350255226?pr=25563) appears to be that we aren't linking with targets.cc which defines hwy::Abort.

Is there a way we can not require `targets.cc` for `VQSortStatic` ?