@Mousius 2024-01-11
> > > The [s390 issue](https://github.com/numpy/numpy/actions/runs/7477479830/job/20350255226?pr=25563) appears to be that we aren't linking with targets.cc which defines hwy::Abort.
> > 
> > 
> > Is there a way we can not require `targets.cc` for `VQSortStatic` ?
> 
> I think it's only required when calling HWY_ABORT/ASSERT/DASSERT, but VQSort does have a few of those, including in VQSortStatic if !VQSORT_ENABLED but VQSortStatic is still called. We could also consider defining hwy::Abort separately in numpy if you really do not want to link with targets.cc, but that seems brittle, right?

Yep, I'd rather not define the Highway function in NumPy, both brittle and weird 😸 I was just curious as up until now we've managed to minimize the size by using `VQSortStatic`, and NumPy isn't going to use much of `targets.cc` other than `Abort` due to having its own dynamic dispatch infrastructure. I wonder if alternatively we could extract `Abort` out of `targets.cc`?