@seiko2plus 2024-01-11
I wonder if there's a way to avoid abort and raising a Python runtime error instead. Which Highway operations require that kind of fatal error? Memory allocations?