@jan-wassenberg 2024-01-12
> It's a bit strange to require header <asm/hwcap.h>, sys/auxv.h should be used instead.

We have hit some systems where one or both are required :/ Highway's CMake includes tooling to check which is available,
you can define TOOLCHAIN_MISS_ASM_HWCAP_H to avoid including that header.

> I wonder if alternatively we could extract Abort out of targets.cc?

I like your idea to pull it out, will do.

> I wonder if there's a way to avoid abort and raising a Python runtime error instead. Which Highway operations require that kind of fatal error? Memory allocations?

Actually allocations do not abort. Mainly it's self-consistency checks or "this is a user/logic error".
Interesting idea with runtime error. How would that be accomplished? We could consider registering an abort handler and hwy::Abort calls that if non-null?