@seiko2plus 2024-01-16
> Is it possible to run benchmarks somewhere to prove this is working?

Currently only PowerPC64 is affected by this patch, here a benchmark but on unstable VM:

<details>
<summary>Power9 benchmark</summary>

```Bash
[100.00%] ··· bench_function_base.Sort.time_sort_worst                                                                                                                                                     186±0.6ms
| Change   | Before [7ed86298]    | After [71e9eb30] <highway_extend_sort_targets>   |   Ratio | Benchmark (Parameter)                                                                    |
|----------|----------------------|--------------------------------------------------|---------|------------------------------------------------------------------------------------------|
| +        | 197±0.4μs            | 801±3μs                                          |    4.07 | bench_function_base.Sort.time_sort('quick', 'int64', ('ordered',))                       |
| +        | 229±0.8μs            | 850±6μs                                          |    3.71 | bench_function_base.Sort.time_sort('quick', 'float64', ('ordered',))                     |
| +        | 302±1μs              | 800±2μs                                          |    2.65 | bench_function_base.Sort.time_sort('quick', 'int64', ('reversed',))                      |
| +        | 338±3μs              | 851±2μs                                          |    2.52 | bench_function_base.Sort.time_sort('quick', 'float64', ('reversed',))                    |
| +        | 195±1μs              | 289±3μs                                          |    1.49 | bench_function_base.Sort.time_sort('quick', 'uint32', ('ordered',))                      |
| +        | 197±0.6μs            | 283±2μs                                          |    1.43 | bench_function_base.Sort.time_sort('quick', 'int32', ('ordered',))                       |
| +        | 661±0.7μs            | 881±1μs                                          |    1.33 | bench_function_base.Sort.time_sort('quick', 'int64', ('sorted_block', 1000))             |
| +        | 732±2μs              | 930±4μs                                          |    1.27 | bench_function_base.Sort.time_sort('quick', 'float64', ('sorted_block', 1000))           |
| +        | 244±1μs              | 307±3μs                                          |    1.26 | bench_function_base.Sort.time_sort('quick', 'float32', ('ordered',))                     |
| +        | 501±10μs             | 594±20μs                                         |    1.19 | bench_function_base.Partition.time_partition('float64', ('sorted_block', 1000), 10)      |
| +        | 505±6μs              | 589±8μs                                          |    1.17 | bench_function_base.Partition.time_partition('float64', ('sorted_block', 1000), 100)     |
| +        | 769±4μs              | 889±3μs                                          |    1.16 | bench_function_base.Sort.time_sort('quick', 'int64', ('sorted_block', 100))              |
| +        | 1.02±0.02ms          | 1.16±0.02ms                                      |    1.14 | bench_function_base.Partition.time_partition('float64', ('sorted_block', 1000), 1000)    |
| +        | 618±4μs              | 697±8μs                                          |    1.13 | bench_function_base.Partition.time_argpartition('float64', ('sorted_block', 1000), 100)  |
| +        | 1.13±0.01ms          | 1.28±0.02ms                                      |    1.13 | bench_function_base.Partition.time_argpartition('float64', ('sorted_block', 1000), 1000) |
| +        | 213±2μs              | 240±4μs                                          |    1.13 | bench_function_base.Partition.time_partition('int16', ('sorted_block', 100), 100)        |
| +        | 760±2μs              | 859±2μs                                          |    1.13 | bench_function_base.Sort.time_sort('quick', 'int64', ('sorted_block', 10))               |
| +        | 605±10μs             | 671±4μs                                          |    1.11 | bench_function_base.Partition.time_partition('float64', ('sorted_block', 100), 10)       |
| +        | 600±5μs              | 667±4μs                                          |    1.11 | bench_function_base.Partition.time_partition('float64', ('sorted_block', 100), 100)      |
| +        | 212±0.7μs            | 234±5μs                                          |    1.11 | bench_function_base.Partition.time_partition('int16', ('sorted_block', 100), 10)         |
| +        | 333±0.4μs            | 370±2μs                                          |    1.11 | bench_function_base.Sort.time_argsort('merge', 'int32', ('sorted_block', 10))            |
| +        | 848±1μs              | 940±9μs                                          |    1.11 | bench_function_base.Sort.time_sort('quick', 'float64', ('sorted_block', 100))            |
| +        | 628±6μs              | 690±10μs                                         |    1.1  | bench_function_base.Partition.time_argpartition('float64', ('sorted_block', 1000), 10)   |
| +        | 561±2μs              | 617±7μs                                          |    1.1  | bench_function_base.Partition.time_partition('int16', ('sorted_block', 10), 100)         |
| +        | 615±7μs              | 662±10μs                                         |    1.08 | bench_function_base.Partition.time_partition('float64', ('sorted_block', 100), 1000)     |
| +        | 236±1μs              | 256±0.5μs                                        |    1.08 | bench_function_base.Sort.time_sort('merge', 'float64', ('sorted_block', 100))            |
| +        | 576±10μs             | 616±6μs                                          |    1.07 | bench_function_base.Partition.time_partition('int16', ('sorted_block', 10), 10)          |
| +        | 582±10μs             | 623±3μs                                          |    1.07 | bench_function_base.Partition.time_partition('int16', ('sorted_block', 10), 1000)        |
| +        | 1.03±0ms             | 1.11±0.01ms                                      |    1.07 | bench_function_base.Sort.time_sort('quick', 'float16', ('random',))                      |
| +        | 950±6μs              | 1.00±0.01ms                                      |    1.06 | bench_function_base.Sort.time_sort('quick', 'float16', ('sorted_block', 10))             |
| +        | 78.3±0.2μs           | 82.5±3μs                                         |    1.05 | bench_function_base.Sort.time_sort('heap', 'float32', ('uniform',))                      |
| -        | 919±6μs              | 873±2μs                                          |    0.95 | bench_function_base.Partition.time_partition('int32', ('sorted_block', 10), 10)          |
| -        | 936±5μs              | 886±3μs                                          |    0.95 | bench_function_base.Partition.time_partition('int64', ('sorted_block', 10), 1000)        |
| -        | 500±7μs              | 472±3μs                                          |    0.94 | bench_function_base.Partition.time_partition('float32', ('sorted_block', 1000), 100)     |
| -        | 940±5μs              | 888±1μs                                          |    0.94 | bench_function_base.Partition.time_partition('int32', ('sorted_block', 10), 1000)        |
| -        | 302±1μs              | 284±0.7μs                                        |    0.94 | bench_function_base.Sort.time_sort('quick', 'int32', ('reversed',))                      |
| -        | 761±10μs             | 692±9μs                                          |    0.91 | bench_function_base.Partition.time_argpartition('float32', ('sorted_block', 1000), 100)  |
| -        | 415±8μs              | 377±2μs                                          |    0.91 | bench_function_base.Sort.time_argsort('merge', 'float16', ('sorted_block', 100))         |
| -        | 201±0.4μs            | 182±2μs                                          |    0.91 | bench_function_base.Sort.time_sort('quick', 'int16', ('ordered',))                       |
| -        | 557±5μs              | 493±10μs                                         |    0.89 | bench_function_base.Partition.time_partition('int64', ('sorted_block', 100), 10)         |
| -        | 546±1μs              | 489±6μs                                          |    0.89 | bench_function_base.Partition.time_partition('int64', ('sorted_block', 100), 100)        |
| -        | 558±10μs             | 482±4μs                                          |    0.86 | bench_function_base.Partition.time_partition('int64', ('sorted_block', 100), 1000)       |
| -        | 364±1μs              | 307±2μs                                          |    0.84 | bench_function_base.Sort.time_sort('quick', 'float32', ('reversed',))                    |
| -        | 186±0.6ms            | 112±0.8ms                                        |    0.6  | bench_function_base.Sort.time_sort_worst                                                 |
| -        | 303±1μs              | 180±0.4μs                                        |    0.59 | bench_function_base.Sort.time_sort('quick', 'int16', ('reversed',))                      |
| -        | 645±2μs              | 325±0.9μs                                        |    0.5  | bench_function_base.Sort.time_sort('quick', 'uint32', ('sorted_block', 1000))            |
| -        | 702±2μs              | 324±2μs                                          |    0.46 | bench_function_base.Sort.time_sort('quick', 'int32', ('sorted_block', 1000))             |
| -        | 807±3μs              | 343±1μs                                          |    0.42 | bench_function_base.Sort.time_sort('quick', 'float32', ('sorted_block', 1000))           |
| -        | 754±1μs              | 316±0.7μs                                        |    0.42 | bench_function_base.Sort.time_sort('quick', 'uint32', ('sorted_block', 100))             |
| -        | 296±30μs             | 122±0.9μs                                        |    0.41 | bench_function_base.Sort.time_argsort('merge', 'int16', ('random',))                     |
| -        | 757±2μs              | 308±0.7μs                                        |    0.41 | bench_function_base.Sort.time_sort('quick', 'uint32', ('sorted_block', 10))              |
| -        | 814±1μs              | 308±1μs                                          |    0.38 | bench_function_base.Sort.time_sort('quick', 'int32', ('sorted_block', 10))               |
| -        | 819±3μs              | 315±0.8μs                                        |    0.38 | bench_function_base.Sort.time_sort('quick', 'int32', ('sorted_block', 100))              |
| -        | 838±1μs              | 308±0.7μs                                        |    0.37 | bench_function_base.Sort.time_sort('quick', 'uint32', ('random',))                       |
| -        | 936±3μs              | 337±0.9μs                                        |    0.36 | bench_function_base.Sort.time_sort('quick', 'float32', ('sorted_block', 100))            |
| -        | 952±3μs              | 328±0.5μs                                        |    0.34 | bench_function_base.Sort.time_sort('quick', 'float32', ('sorted_block', 10))             |
| -        | 906±2μs              | 309±1μs                                          |    0.34 | bench_function_base.Sort.time_sort('quick', 'int32', ('random',))                        |
| -        | 1.06±0ms             | 327±1μs                                          |    0.31 | bench_function_base.Sort.time_sort('quick', 'float32', ('random',))                      |
| -        | 738±2μs              | 204±0.8μs                                        |    0.28 | bench_function_base.Sort.time_sort('quick', 'int16', ('sorted_block', 1000))             |
| -        | 848±3μs              | 201±4μs                                          |    0.24 | bench_function_base.Sort.time_sort('quick', 'int16', ('sorted_block', 10))               |
| -        | 203±3μs              | 45.9±0.2μs                                       |    0.23 | bench_function_base.Sort.time_sort('quick', 'float64', ('uniform',))                     |
| -        | 858±1μs              | 194±1μs                                          |    0.23 | bench_function_base.Sort.time_sort('quick', 'int16', ('sorted_block', 100))              |
| -        | 952±5μs              | 196±0.7μs                                        |    0.21 | bench_function_base.Sort.time_sort('quick', 'int16', ('random',))                        |
| -        | 211±0.5μs            | 31.4±0.05μs                                      |    0.15 | bench_function_base.Sort.time_sort('quick', 'float32', ('uniform',))                     |
| -        | 159±2μs              | 20.3±0.2μs                                       |    0.13 | bench_function_base.Sort.time_sort('quick', 'int64', ('uniform',))                       |
| -        | 166±2μs              | 14.3±0.1μs                                       |    0.09 | bench_function_base.Sort.time_sort('quick', 'int32', ('uniform',))                       |
| -        | 157±1μs              | 14.4±0.1μs                                       |    0.09 | bench_function_base.Sort.time_sort('quick', 'uint32', ('uniform',))                      |
| -        | 175±2μs              | 11.2±0.2μs                                       |    0.06 | bench_function_base.Sort.time_sort('quick', 'int16', ('uniform',))                  
```
</details>

>  Is there any effect on benchmarks on other architectures (there shouldn't be, just looking for confirmation)?

This patch only limited to PowerPC64.